Desktop/yzkcb copy/requirements.txt
# Flask Backend - REST API
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0

# Web UI - Remove Gradio after migration
# gradio>=4.0.0  # DEPRECATED - Remove after frontend migration complete

# LLM & LangChain
langchain>=0.0.350
langchain-community>=0.0.20
langchain-core>=0.1.20

# Embeddings & Vector Store (GPU-accelerated)
sentence-transformers==2.7.0  # Stable version compatible with Python 3.11
faiss-cpu>=1.7.4  # Note: GPU acceleration via CUDA/PyTorch
torch>=2.5.1
torchvision>=0.20.1
sentencepiece>=0.2.0  # Required by sentence-transformers

# Document Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0

# RAG & Retrieval
chromadb>=0.4.0
rank-bm25>=0.2.2

# NLP & Text Processing
nltk>=3.8.0
spacy>=3.7.0
textstat>=0.7.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Database
pymongo>=4.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
requests>=2.31.0

# Security
cryptography>=41.0.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Database
sqlalchemy>=2.0.0

# API
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI
gradio>=4.0.0

# Markdown processing
markdown>=3.5.0
